Add patient module moved from pages to apps module
