+++ apps/modernization-ui/src/apps/patient/add/DataEntrySideNav/DataEntrySideNav.scss
