+++ apps/modernization-ui/src/apps/patient/add/personalDetails/PersonalDetails.scss
