--- apps/modernization-ui/src/pages/patient/add/index.ts
@@ -1,1 +1,0 @@
-export * from './NewPatientEntry';
